Manual rebase 2

+++ app/controllers/offsprings_controller.rb
@@ -1,34 +1,43 @@
 # Offsprings are dependent on users and are the ones that get assigned to shifts
 # They can't be assigned themselves, and need the process to be done by the user
 class OffspringsController < ApplicationController
+  SCOPE = "activerecord.errors.controllers.offspring".freeze # Necessary to acces locales file
   def show
   end
 
   def index
   end
-  
+
   def new
     @offspring = current_user.offsprings.build
   end
 
   def create
-    @offspring = current_user.offsprings.build(offsprings_params)
-    if !@offspring.primary_first?
-      flash[:warning] = "Sólo puede añadir a niños de 1º de Primaria"
-      redirect_to static_pages_intructions_path
-      return
-    elsif @offspring.save
-      flash[:success] = "Niño añadido"
+    if admin_allows_changes?
+      @offspring = current_user.offsprings.build(offsprings_params)
+      if !@offspring.primary_first?
+        flash[:warning] = "Sólo puede añadir a niños de 1º de Primaria"
+        redirect_to static_pages_intructions_path
+        return
+      elsif @offspring.save
+        flash[:success] = "Niño añadido"
+      else
+        flash[:danger] = "No se pudo añadir al niño"
+      end
     else
-      flash[:danger] = "No se pudo añadir al niño"
+      flash[:alert] = I18n.t("admin_locked_create", scope: SCOPE)
     end
     redirect_to root_path
   end
 
   def destroy
-    offspring = Offspring.find_by_id(params[:id])
-    offspring.destroy
-    flash[:success] = "Niño borrado"
+    if admin_allows_changes?
+      offspring = Offspring.find_by_id(params[:id])
+      offspring.destroy
+      flash[:success] = "Niño borrado"
+    else
+      flash[:alert] = I18n.t("admin_locked_destroy", scope: SCOPE)
+    end
     redirect_to request.referer || root_url
   end
 
@@ -36,5 +45,9 @@ class OffspringsController < ApplicationController
 
   def offsprings_params
     params.require(:offspring).permit(:first_name, :last_name, :grade)
+  end
+
+  def admin_allows_changes?
+    !ApplicationHelper.status_lock?
   end
 end
